added SettingsState

+++ Engine/HGauge.cpp
@@ -14,7 +14,7 @@ namespace Engine
 	void	HGauge::draw(Engine::IRender* render)
 	{
 		this->_filledImg->setPosition(this->_globalPosition);
-		this->_filledImg->setSize(Ultra::Vector2D<long unsigned int>(this->getWidth() * this->_percent / 100, this->getHeight()));
+		this->_filledImg->setSize(Ultra::Vector2D<size_t>(this->getWidth() * this->_percent / 100, this->getHeight()));
 		this->_filledImg->draw(render);
 		this->_bgImg->setPosition(this->_globalPosition);
 		this->_bgImg->setSize(this->_size);
